pyproject.toml
[project]
name = "gen-ai-trabajo"
version = "0.1.0"
description = "Fake news detection using GenAI and LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"langchain-core>=0.2.0",
	"langchain-openai>=0.1.0",
	"langgraph>=0.2.0",
	"openai>=1.40.0",
	"python-dotenv>=1.0.1",
	"tavily-python>=0.3.3",
	"httpx>=0.27.0",
	"beautifulsoup4>=4.12.0",
	"pandas>=3.0.3",
	"sentence-transformers>=5.5.0",
	"qdrant-client>=1.18.0",
]
